Refactor ZIP precheck into dedicated module
